@@ -1,6 +1,7 @@
 {
   "name": "moviedb-project",
-  "version": "0.0.0",
+  "version": "1.0.0",
+  "main": "index.js",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -11,28 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "0.802.0",
-    "@angular/animations": "7.2.15",
+    "@angular/animations": "8.2.1",
     "@angular/cdk": "7.3.7",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
+    "@angular/common": "~8.2.1",
+    "@angular/compiler": "~8.2.1",
+    "@angular/core": "~8.2.1",
+    "@angular/fire": "5.2.1",
     "@angular/flex-layout": "8.0.0-beta.26",
-    "@angular/forms": "~7.2.0",
+    "@angular/forms": "~8.2.1",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/platform-browser": "~8.2.1",
+    "@angular/platform-browser-dynamic": "~8.2.1",
+    "@angular/router": "~8.2.1",
+    "angularfire2": "5.2.1",
     "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
+    "firebase": "6.3.5",
+    "rxjs": "~6.5.2",
     "rxjs-compat": "6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "typescript": "3.5.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "8.2.0",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "0.12.4",
+    "@angular/cli": "8.2.1",
+    "@angular/compiler-cli": "~8.2.1",
+    "@angular/language-service": "~8.2.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -47,7 +52,16 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2",
     "xxxxx": "1.0.3"
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ivdim1992/angular-moviedb-project.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ivdim1992/angular-moviedb-project/issues"
+  },
+  "homepage": "https://github.com/ivdim1992/angular-moviedb-project#readme"
 }
