@@ -1,1 +1,1 @@
-export * from './endpoints';
+export * from './urls.enums';
