@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyB7AKvsr0R_bg6orCvpXmpjXBL43PMV2J0',
+    authDomain: 'moviedb-217e9.firebaseapp.com',
+    databaseURL: 'https://moviedb-217e9.firebaseio.com',
+    projectId: 'moviedb-217e9',
+    storageBucket: 'moviedb-217e9.appspot.com',
+    messagingSenderId: '679955006888'
+  }
 };
