@@ -3,7 +3,15 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyB7AKvsr0R_bg6orCvpXmpjXBL43PMV2J0',
+    authDomain: 'moviedb-217e9.firebaseapp.com',
+    databaseURL: 'https://moviedb-217e9.firebaseio.com',
+    projectId: 'moviedb-217e9',
+    storageBucket: 'moviedb-217e9.appspot.com',
+    messagingSenderId: '679955006888'
+  }
 };
 
 /*
